se cambio punto para que acepte coordenadas float, se integro k-nearest

+++ Rtree_core/Point.h
@@ -7,12 +7,12 @@ class Point
 {
 public:
     Point();
-    Point(int, int);
-    int getX();
-    int getY();
+    Point(float, float);
+    float getX();
+    float getY();
 private: 
-    int cx;
-    int cy;
+    float cx;
+    float cy;
 //
 };
